fix: keep config.rs under the line limit and drop the manual version bump
- Split the admin-UI constructors into src/admin_config.rs and the
  provider defaults into src/config_defaults.rs (both re-exported from
  crate::config), so src/config.rs is back under the 1000-line limit
  that CI's check-file-size.rs enforces.
- Revert Cargo.toml to 0.24.0: this repo forbids manual version edits
  and derives the release from the changelog fragment, which now
  carries the required 'bump: minor' frontmatter.
