+++ changelog.d/20260807_190000_issue_50_admin_ui.md
@@ -1,1 +1,7 @@
+---
+bump: minor
+---
+
+### Added
+
 - Added an opt-in admin UI ([#50](https://github.com/link-assistant/router/issues/50)): a minimal React console built on **Chakra UI v3** (Horizon UI as a visual reference only, not a v2 fork), built with Vite and embedded in the binary with `rust-embed`, served by the router on a **separate listener that does not exist unless you give it a port** — `--admin-port` / `ADMIN_PORT`, bound independently of the proxy via `--admin-host` / `ADMIN_HOST` (default `127.0.0.1`), so the proxy can face the network while the console stays on loopback. The admin credential is either provisioned at deploy time (`TOKEN_ADMIN_KEY`, which closes bootstrap from the start) or **claimed by the first browser visit** through a two-phase handshake that cannot brick the deployment: `POST /api/admin/bootstrap` mints a candidate token and a `claim_id` while persisting nothing and authorising nothing, the client writes the token to `localStorage` and **reads it back**, and only `POST /api/admin/bootstrap/confirm` — authenticated with that freshly stored token itself — activates the credential and closes bootstrap. An unconfirmed candidate expires after `--admin-claim-ttl-secs` (default 120s) leaving the system unclaimed and the mint retryable; only one candidate is outstanding at a time so the first to *confirm* wins and the loser's candidate is discarded; and when the storage read-back fails the client deliberately does **not** confirm, showing the token for manual copy and leaving bootstrap open. Only the SHA-256 digest of the active credential is persisted (`<data-dir>/admin-claim.json`). The console lists tokens (id, label, issued, expires, used/max requests, revoked), issues them (label, TTL, optional request cap, optional account pin) showing each value exactly once, revokes behind a confirmation dialog, rotates the admin credential, and reports read-only status (`/v1/accounts`, `/v1/usage`, doctor-style credential state); `doctor` now prints `admin_ui` and `admin_credential`. The `localStorage` exposure risk is documented in `docs/use-cases/admin-ui.md`. Covered by `tests/admin_ui_test.rs` (12 passed) driving the real admin router end to end plus 10 unit tests in `src/admin.rs`, including that a mint alone neither authorises nor closes bootstrap, that an expired candidate leaves the system unclaimed, and that only the first confirmer wins.
