+++ src/lib.rs
@@ -8,6 +8,7 @@ pub mod accounts;
 pub mod activitypub;
 pub mod admin;
 pub mod admin_api;
+pub mod admin_config;
 pub mod admin_ui;
 pub mod anthropic_bridge;
 pub mod anthropic_stream;
@@ -16,6 +17,7 @@ pub mod audit;
 pub mod claude_identity;
 pub mod cli;
 pub mod config;
+pub mod config_defaults;
 pub mod crater;
 pub mod gemini;
 pub mod gonka;
